requirements.txt
streamlit==1.18.0
google-generativeai==0.1.0
